ONT-671 update governance contract (#265)
* ONT-671 update governance contract

* add some err checks

* change vbft config from uint64 to uint32

+++ consensus/vbft/utils.go
@@ -140,14 +140,14 @@ func GetVbftConfigInfo() (*config.VBFTConfig, error) {
 		return nil, err
 	}
 	chainconfig := &config.VBFTConfig{
-		N:                    cfg.N,
-		C:                    cfg.C,
-		K:                    cfg.K,
-		L:                    cfg.L,
-		BlockMsgDelay:        cfg.BlockMsgDelay,
-		HashMsgDelay:         cfg.HashMsgDelay,
-		PeerHandshakeTimeout: cfg.PeerHandshakeTimeout,
-		MaxBlockChangeView:   cfg.MaxBlockChangeView,
+		N:                    uint32(cfg.N),
+		C:                    uint32(cfg.C),
+		K:                    uint32(cfg.K),
+		L:                    uint32(cfg.L),
+		BlockMsgDelay:        uint32(cfg.BlockMsgDelay),
+		HashMsgDelay:         uint32(cfg.HashMsgDelay),
+		PeerHandshakeTimeout: uint32(cfg.PeerHandshakeTimeout),
+		MaxBlockChangeView:   uint32(cfg.MaxBlockChangeView),
 	}
 	return chainconfig, nil
 }
